Add files via upload
Your own Unity project should have its own Assets folder, which is where just about everything you've done should be, unless you're using some other package or something fancy.

Assets folder. I think the SampleScene in the Scenes folder will have the GameObjects present from the start (the PlayerCharacter, Master, RoomGenerator, and Main Camera as a child of the player char). If not, could probably adjust to not rely on that.


Generates a random room, allows the player to move with WASD or arrow keys, and if the player moves to the exit door on the opposite side from which they started, a new random room from the pool of possible rooms will be generated. 

Has two types/methods for generation, which you can switch between in MasterScript's Start() function (commenting out the 2 lines you aren't using). 
One uses a set of arrays of ints, present in RoomGenerationScript to create a room by looping through the array and placing a tile based on the int (1 for floor, 2 for wall, etc).
The other uses Unity's Tilemap system, and RoomGenerationScript has an array of GameObjects within its component in the Inspector for RoomGenerator. Those GameObjects are Prefabs of Rooms built using the Tilemap system, which you can see in Assets at the bottom of the Unity project in your own editor.

Let me know if you have issues figuring out the file system, getting it working, or questions about the code.