Basic player walk animations and movement controls
4 direction basic walk cycle animations and keyboard controls.
Should be pretty much self-contained, just drag the folder into your Assets. Includes a prefab of the player character in the folder as well. 
The main camera is a child of the player character (to keep the character centered as it moves), so that bit might require some adjustment since you probably have a main camera in your scene already.